Modif .gitignore architecture restructuration requirement.txt knn model
