--- src/modeling/modeling.py
@@ -1,30 +1,0 @@
-
-import numpy as np
-import pandas as pd
-
-from sklearn.neighbors import KNeighborsRegressor
-
- 
-
-
-def impute_knn(df):
-    
-    ''' inputs: pandas df containing feature matrix '''
-    ''' outputs: dataframe with NaN imputed '''
-    # imputation with KNN unsupervised method
- 
-    # separate dataframe into numerical/categorical
-    ldf = df.select_dtypes(include=[np.number])           # select numerical columns in df
-    ldf_putaside = df.select_dtypes(exclude=[np.number])  # select categorical columns in df
-    # define columns w/ and w/o missing data
-    cols_nan = ldf.columns[ldf.isna().any()].tolist()         # columns w/ nan 
-    cols_no_nan = ldf.columns.difference(cols_nan).values     # columns w/o nan 
-
-    for col in cols_nan:                
-        imp_test = ldf[ldf[col].isna()]   # indicies which have missing data will become our test set
-        imp_train = ldf.dropna()          # all indicies which which have no missing data 
-        model = KNeighborsRegressor(n_neighbors=5)  # KNR Unsupervised Approach
-        knr = model.fit(imp_train[cols_no_nan], imp_train[col])
-        ldf.loc[df[col].isna(), col] = knr.predict(imp_test[cols_no_nan])
-    
-    return pd.concat([ldf,ldf_putaside],axis=1)
